requirements.txt
plaid-python==19.0.0
python-dotenv==1.0.1
SQLAlchemy==2.0.32
pandas==2.2.2
pydantic==2.8.2
openai==1.40.2
tiktoken==0.7.0
streamlit==1.37.1
tabulate==0.9.0
rich==13.7.1